Move CommandParser to command package
